Extract commands to their own modules
