+++ lib/makina/cli.ex
@@ -1,16 +1,20 @@
 defmodule Makina.Cli do
-  alias Owl.IO
+  alias Makina.Cli.Commands
 
-  alias Makina.SSH
+  @commands %{
+    help: Commands.Help,
+    init: Commands.Init,
+    test: Commands.Test
+  }
 
   def start(_, [:test]) do
     {:ok, self()}
   end
 
   def start(_, _args) do
-    {command, _arguments, _options} = parse_command(Burrito.Util.Args.argv())
+    {command, arguments, options} = parse_command(Burrito.Util.Args.argv())
 
-    case command(command) do
+    case command(command, arguments, options) do
       :ok -> System.halt(0)
       :error -> System.halt(-1)
     end
@@ -18,97 +22,11 @@ defmodule Makina.Cli do
     :ok
   end
 
-  def command(command, arguments \\ [], options \\ [])
+  def command(command, arguments \\ [], options \\ []) do
+    module = Map.get(@commands, command, Commands.Help)
 
-  def command(:help, _arguments, _options) do
-    message = """
-    Makina
-    Paolo Marino
-    A simple application manager for self-hosted environments.
-
-    Usage: Makina <COMMAND> [OPTIONS]
-
-    Arguments:
-    help    Prints the general help message
-    init    Initializes the current repository with an empty Makinafile
-    test    Tests whether nodes configured inside the Makinafile can be successully reached
-
-    Options:
-     -h, --help       Print help
-    """
-
-    IO.puts(message)
-
-    :ok
+    module.exec(arguments, options)
   end
-
-  def command(:init, arguments, _options) do
-    destination_path = List.first(arguments, File.cwd!())
-    file = Path.join(destination_path, "Makinafile.exs")
-
-    if File.exists?(file) do
-      IO.puts("File already exists, skipping!")
-    else
-      File.write(file, """
-      # Hello
-      """)
-
-      IO.puts("Makinafile created at #{Path.relative_to_cwd(file)} ✅")
-    end
-
-    {:ok, ""}
-  end
-
-  def command(:test, _arguments, _options) do
-    makinafile = Path.join(File.cwd!(), "Makinafile.exs")
-
-    Makina.File.compile_build_file(makinafile)
-    servers = Makina.File.fetch_servers()
-
-    Owl.ProgressBar.start(
-      id: :server_tests,
-      label: "Testing servers",
-      total: Enum.count(servers)
-    )
-
-    results =
-      for server <- servers do
-        response =
-          SSH.connect(
-            server.host,
-            user: server.user,
-            password: server.password,
-            port: server.port
-          )
-
-        case response do
-          {:ok, conn_ref} ->
-            SSH.disconnect(conn_ref)
-            Owl.ProgressBar.inc(id: :server_tests)
-
-            {:ok, server}
-
-          {:error, _reason} ->
-            Owl.ProgressBar.inc(id: :server_tests)
-
-            {:error, server}
-        end
-      end
-
-    Owl.LiveScreen.await_render()
-
-    if test_with_errors?(results) do
-      IO.puts(Owl.Data.tag("Cannot reach some servers", :red))
-
-      :error
-    else
-      IO.puts("All servers are reachable ✅")
-
-      :ok
-    end
-  end
-
-  def command(_cmd, arguments, options), do: command(:help, arguments, options)
 
   @doc """
   Parses a list of cli args and returns a tuple in the form of `{command, arguments, options}`
@@ -129,7 +47,4 @@ defmodule Makina.Cli do
   end
 
   defp options_for(_cmd), do: []
-
-  defp test_with_errors?(results),
-    do: Enum.any?(results, fn r -> elem(r, 0) == :error end)
 end
