+++ lib/makina/cli/command.ex
@@ -1,0 +1,4 @@
+defmodule Makina.Cli.Command do
+  @callback exec(arguments :: list(), options :: list()) :: :ok | :error
+  @callback options() :: list(keyword())
+end
